Put back compilation and load into the Lemmatizer. Added tests to check this function
